+++ sdk/containerregistry/azure-containerregistry/tests/test_container_repository_client.py
@@ -11,7 +11,7 @@ from devtools_testutils import AzureTestCase
 from azure.containerregistry import (
     ContainerRepositoryClient,
     ContainerRegistryClient,
-    ContentPermissions,
+    ContentProperties,
     DeletedRepositoryResult,
     RepositoryProperties,
     RegistryArtifactOrderBy,
@@ -46,17 +46,17 @@ class TestContainerRepositoryClient(ContainerRegistryTestClass):
 
     @acr_preparer()
     def test_delete_tag_does_not_exist(self, containerregistry_endpoint):
-        client = self.create_repository_client(containerregistry_endpoint, HELLO_WORLD)
+        client = self.create_repository_client(containerregistry_endpoint, "DOES_NOT_EXIST123")
 
         with pytest.raises(ResourceNotFoundError):
-            client.delete_tag(TO_BE_DELETED)
+            client.delete_tag("DOESNOTEXIST123")
 
     @acr_preparer()
     def test_get_properties(self, containerregistry_endpoint):
         repo_client = self.create_repository_client(containerregistry_endpoint, HELLO_WORLD)
 
         properties = repo_client.get_properties()
-        assert isinstance(properties.content_permissions, ContentPermissions)
+        assert isinstance(properties.writeable_properties, ContentProperties)
 
     @acr_preparer()
     def test_get_tag(self, containerregistry_endpoint):
@@ -203,11 +203,11 @@ class TestContainerRepositoryClient(ContainerRegistryTestClass):
         client = self.create_repository_client(containerregistry_endpoint, repository)
 
         tag_props = client.get_tag_properties(tag_identifier)
-        permissions = tag_props.content_permissions
+        permissions = tag_props.writeable_properties
 
         received = client.set_tag_properties(
             tag_identifier,
-            ContentPermissions(
+            ContentProperties(
                 can_delete=False,
                 can_list=False,
                 can_read=False,
@@ -215,14 +215,14 @@ class TestContainerRepositoryClient(ContainerRegistryTestClass):
             ),
         )
 
-        assert not received.content_permissions.can_write
-        assert not received.content_permissions.can_read
-        assert not received.content_permissions.can_list
-        assert not received.content_permissions.can_delete
+        assert not received.writeable_properties.can_write
+        assert not received.writeable_properties.can_read
+        assert not received.writeable_properties.can_list
+        assert not received.writeable_properties.can_delete
 
         client.set_tag_properties(
             tag_identifier,
-            ContentPermissions(
+            ContentProperties(
                 can_delete=True,
                 can_list=True,
                 can_read=True,
@@ -235,7 +235,7 @@ class TestContainerRepositoryClient(ContainerRegistryTestClass):
         client = self.create_repository_client(containerregistry_endpoint, self.get_resource_name("repo"))
 
         with pytest.raises(ResourceNotFoundError):
-            client.set_tag_properties(DOES_NOT_EXIST, ContentPermissions(can_delete=False))
+            client.set_tag_properties(DOES_NOT_EXIST, ContentProperties(can_delete=False))
 
     @acr_preparer()
     def test_set_manifest_properties(self, containerregistry_endpoint, containerregistry_resource_group):
@@ -246,11 +246,11 @@ class TestContainerRepositoryClient(ContainerRegistryTestClass):
         client = self.create_repository_client(containerregistry_endpoint, repository)
 
         for artifact in client.list_registry_artifacts():
-            permissions = artifact.content_permissions
+            permissions = artifact.writeable_properties
 
             received_permissions = client.set_manifest_properties(
                 artifact.digest,
-                ContentPermissions(
+                ContentProperties(
                     can_delete=False,
                     can_list=False,
                     can_read=False,
@@ -258,15 +258,15 @@ class TestContainerRepositoryClient(ContainerRegistryTestClass):
                 ),
             )
 
-            assert not received_permissions.content_permissions.can_delete
-            assert not received_permissions.content_permissions.can_read
-            assert not received_permissions.content_permissions.can_list
-            assert not received_permissions.content_permissions.can_write
+            assert not received_permissions.writeable_properties.can_delete
+            assert not received_permissions.writeable_properties.can_read
+            assert not received_permissions.writeable_properties.can_list
+            assert not received_permissions.writeable_properties.can_write
 
             # Reset and delete
             client.set_manifest_properties(
                 artifact.digest,
-                ContentPermissions(
+                ContentProperties(
                     can_delete=True,
                     can_list=True,
                     can_read=True,
@@ -281,7 +281,7 @@ class TestContainerRepositoryClient(ContainerRegistryTestClass):
         client = self.create_repository_client(containerregistry_endpoint, self.get_resource_name("repo"))
 
         with pytest.raises(ResourceNotFoundError):
-            client.set_manifest_properties("sha256:abcdef", ContentPermissions(can_delete=False))
+            client.set_manifest_properties("sha256:abcdef", ContentProperties(can_delete=False))
 
     @acr_preparer()
     def test_delete_repository(self, containerregistry_endpoint, containerregistry_resource_group):